SwinMM/requirements.txt
apex==0.9.10dev
batchgenerators==0.25
caffe2==0.8.1
colorama==0.4.6
Flask==2.3.2
gevent==23.7.0
gorilla==0.4.0
hypothesis==6.81.2
lz4==4.3.2
monai==1.2.0
nibabel==5.1.0
numba==0.57.1
numpy==1.25.1
pssh==2.3.1
ptvsd==4.3.2
pydantic==2.0.3
pylzma==0.5.0
pytest==7.4.0
redis==4.6.0
Requests==2.31.0
scipy==1.11.1
setuptools==65.6.3
setuptools_rust==1.6.0
tensorboardX==2.6.1
tensorflow_datasets==4.9.2
timm==0.9.2
torchvision==0.15.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.7.1
urllib3==1.26.15
xmlrunner==1.7.7
xxhash==3.2.0
